requirements.txt
asgiref==3.8.1
bleach==6.2.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
cryptography==44.0.1
defusedxml==0.7.1
Django==5.1.4
django-cors-headers==4.7.0
django-filter==24.3
django-jazzmin==3.0.1
django-svg-image-form-field==1.0.1
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
djoser==2.3.1
drf-yasg==1.21.8
flake8==7.1.1
gunicorn==23.0.0
idna==3.10
inflection==0.5.1
Markdown==3.5.2
martor==1.6.45
mccabe==0.7.0
oauthlib==3.2.2
packaging==24.2
pillow==11.1.0
pycodestyle==2.12.1
pycparser==2.22
pyflakes==3.2.0
PyJWT==2.10.1
python3-openid==3.2.0
pytz==2025.1
PyYAML==6.0.2
requests==2.32.3
requests-oauthlib==2.0.0
social-auth-app-django==5.4.3
social-auth-core==4.5.6
sqlparse==0.5.3
stripe==11.4.1
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.3.0
webencodings==0.5.1
